+++ src/main/java/pl/arabowski/CSDTSCoding_Assignment/service/EventService.java
@@ -1,6 +1,7 @@
 package pl.arabowski.CSDTSCoding_Assignment.service;
 
 import java.io.IOException;
+import java.math.BigInteger;
 
 import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.databind.JsonMappingException;
@@ -11,7 +12,7 @@ public interface EventService {
 	
 	public Event convertToObject(String jsonFile) throws JsonParseException, JsonMappingException, IOException;
 	
-	public long countDuration(long startTime, long finishTime);
+	public BigInteger countDuration(BigInteger startTime, BigInteger finishTime);
 	
 	public Iterable<?> readFromFile (String filePath);
